Adds basic support for the Stream Deck Mini
The Streamdeck mini has a different USB PID (and obviously number of
buttons and layout) than the original Streamdeck.  The changes here
enable basic device recognition and keypress detection.  Due to some
underlying protocol changes, updating the button images doesn't
currently work on the mini.  (Issue 56 on Lange/node-elgato-stream-deck
has some discussion of the details.)

StreadDeckMini is implemented as a subclass of StreamDeck.  Instead of
just a single list, DeviceManager.enumerate() now returns a tuple
consisting of a list of original StreamDeck objects and a list of
StreamDeckMini objects, which is probably not the most efficient way to
do it, but it works.

Alterations to example.py are made to draw on the additional
information, including a new method on the parent class that returns a
friendly string naming the model associated with the current object.