Refactor _read_key_states into _read_control_states
The updated internal function returns a dictionary with essentially the same data
as before, but the device implementation itself is handling report decoding.

While it is possible a future report might contain multiple types of return data,
current reports do not on the only device capable of multiple control types.

This commit should bring #108 closer to merge by implementing the first phase of
https://github.com/abcminiuser/python-elgato-streamdeck/pull/108#issuecomment-1368328529